requirements.txt
python-binance==1.0.19
requests==2.31.0